Update MPI_Demo1.c
We would like to write a program to create: n processes with the ranks: 0, 1, 2, 3,…(n-1). The processes with the rank 1, 2, 3, ….(n-1) sends a message to the process with the rank: 0. This process receives all of these messages and displays them on the screen. This process also prints a message for itself. In other words:

If I am process1, process2, … or process n-1
            Send my message to process0.
else if I am process0
            Print my message.
            Receive messages from other processes and print them.